Rewrite CS code, add proper testing, and Latin parsing support.
This is a BAD thing to be doing, but, I got carried away
with myself, and so, given the personal nature of this project,
I am OK with this, vs spending the large amount of time splitting
all of this up. Please don't judge- I won't be getting into the habit.

This does four main things:

1. I split up a lot of CoffeeScript code into smaller and
more succinct modules.

2. Fleshed out the testing strategy (for the CoffeeScript layer),
and added unit and integration tests for various modules and
sets of modules, as well as added more system level testing for
the library and cli portion. It is not perfect, but a better
ground to work on. This includes flushing out the system
test fixtures and example files.

3. Added support for latin text matching (for fun). :-)

4. Attempt to unify the CLI APIs for all 3 languages, including
the `-s` and `-m` options.

Note: This also adds a bunch of other little things and fixes,
mainly to do with wording, etc.