Type casting modules is not really a good idea.
It was a fun thing to try, but in the end it ends
up masking type breakages and can provide false positives.

However we can set an interface check for the main exported library,
which applies downward to each module's function/property types that
typescript automatically picks up.

Oh Haskell... <3